Fix segfault and window closing bug
The segfault was caused by the use of double backspace characters in
BoolButton.cpp, NumberButton.cpp, and Dropdown.cpp when accessing the
fonts. Fixing it was as simple as changing it to a single forward slash.

There is also a bug that prevented you from closing the window due to
the use of deprecated SDL2 window closing events. Replacing it with the
modern version fixed it completely.