Add installation file for Dependencies
Add bash script to automatically install dependencies. As a result,
delete Dependencies/ from the git project entirely. Update .gitignore to
ignore those.

In addition to this, add a compile.bash file that automatically compiles
the project into executable. As a result, change the path for fonts in
some files and remove all .dll from root.